fixing compile errors

+++ cynchMain/client/src/com/muddyhorse/cynch/ProgressListener.java
@@ -7,7 +7,7 @@ public interface ProgressListener
 {
     public void starting(Operation op);
 
-    public void progress(String name, String desc, int amount, int total) throws InterruptedException;
+    public void progress(String name, String desc, long amount, long total) throws InterruptedException;
 
     public void finished(Operation op, boolean success);
 }
